Move API request to a helper
