+++ services/enrichment/fields/budget/src/lib/processBudgetItem.js
@@ -1,9 +1,8 @@
 // @flow
-import request from 'request-promise-native';
 import sanitizeBudgetItem from '@eubfr/lib/budgetFormatter';
 import type { BudgetItem } from '@eubfr/types';
 
-const precision = 100; // round converted value to the nearest 100
+import { getEuroValue } from './getEuroValue';
 
 // List of currencies that can be translated into EUR
 // Manually extracted form ECB's API
@@ -55,9 +54,6 @@ const availableCurrencies = [
   'ZAR',
 ];
 
-const padDate = (date: number): number | string =>
-  date < 10 ? `0${date}` : date;
-
 export const processBudgetItem = async (
   inputBudgetItem: BudgetItem,
   projectEnd: string,
@@ -78,66 +74,14 @@ export const processBudgetItem = async (
   ) {
     const projectEndDate = new Date(projectEnd);
 
-    // Least precision: year
-    let period = projectEndDate.getFullYear();
-    let apiPeriod = 'A';
+    const euroValue = await getEuroValue(
+      inputBudgetItem,
+      projectEndDate,
+      projectEndPrecision
+    );
 
-    // Best precision: day
-    if (projectEndPrecision === 'day') {
-      const month = projectEndDate.getMonth() + 1;
-      const day = projectEndDate.getDate();
-      period = `${period}-${padDate(month)}-${padDate(day)}`;
-      apiPeriod = 'D';
-    }
-
-    // If not a day, try at least: month
-    if (projectEndPrecision === 'month') {
-      const month = projectEndDate.getMonth() + 1;
-      period = `${period}-${padDate(month)}`;
-      apiPeriod = 'M';
-    }
-
-    // "EXR/A.[currency].EUR.SP00.A" corresponds to the key of a dataset
-    // The keys can be found here: http://sdw.ecb.europa.eu/browseTable.do?node=1495
-    // API info: https://sdw-wsrest.ecb.europa.eu/web/generator/index.html
-    const url = `https://sdw-wsrest.ecb.europa.eu/service/data/EXR/${apiPeriod}.${
-      inputBudgetItem.currency
-    }.EUR.SP00.A`;
-
-    let results;
-    const qs = {
-      startPeriod: period,
-      endPeriod: period,
-      detail: 'dataonly',
-    };
-
-    try {
-      results = await request.get({
-        url,
-        qs,
-        json: true,
-        headers: {
-          Accept: 'application/vnd.sdmx.data+json;version=1.0.0-wd',
-        },
-      });
-    } catch (e) {
-      console.error('error', url, qs, e);
-      return inputBudgetItem; // budget not enriched
-    }
-
-    if (
-      results &&
-      results.dataSets &&
-      results.dataSets[0] &&
-      results.dataSets[0].series &&
-      results.dataSets[0].series['0:0:0:0:0'] &&
-      results.dataSets[0].series['0:0:0:0:0'].observations &&
-      results.dataSets[0].series['0:0:0:0:0'].observations[0]
-    ) {
-      const exr = results.dataSets[0].series['0:0:0:0:0'].observations['0'][0];
-      const euroValue =
-        Math.ceil(inputBudgetItem.value / exr / precision) * precision;
-
+    // If useful value and if changed, i.e. non-original value
+    if (euroValue && euroValue.value !== inputBudgetItem.value) {
       const formattedEuroBudget = sanitizeBudgetItem({
         value: euroValue,
         currency: 'EUR',
